texture_extraction/requirements.txt
imageio>=2.0
imageio-ffmpeg
numpy
Pillow
Wand
